+++ downpour_sgd.py
@@ -1,9 +1,10 @@
 import gevent
+from utils import Messages, send_message, squash_model, set_params
 import torch
-from utils import Messages, squash_model, ACTION_CODES, CODE_ACTIONS, send_message
-import torch.distributed as dist
-
+import time
 import logging
+from models.mnist import Net
+import torch.distributed as dist
 
 logging.basicConfig(
     format='%(asctime)s %(levelname)-8s %(message)s',
@@ -12,30 +13,26 @@ logging.basicConfig(
 
 _LOGGER = logging.getLogger(__name__)
 
-class ParameterServer():
+DEFAULT_LEARNING_RATE = 0.005
 
-    def __init__(self, learning_rate, model, random_seed=42, rank=0, size=0):
-        """__init__
 
-        :param learning_rate:
-        :param params: a dict of str -> pytorch tensor that represents the gradients
-        :param random_seed:
-        """
-        _LOGGER.info("Creating ParameterServer with LR: {}".format(learning_rate))
+
+class DownpourSGD():
+
+    def __init__(self, learning_rate, model):
         self.learning_rate = learning_rate
+        self.request_frequency = 5 # request every 5 self.runs (including 0)
+        self.model = model
         _LOGGER.info("Setting m_parameter")
         self.m_parameter = torch.zeros(squash_model(model).numel() + 1)
 
     def receive(self, message, parameter):
         print("Processing message: {}".format(message))
-        if message == 'ParameterRequest':
-            send_message('ParameterUpdate', self.parameters, dst=1)    
-
-        elif message == 'GradientUpdate':
-            self.parameters -= self.learning_rate * gradient
+        if message == 'ParamaterUpdate':
+            set_params(self.model, parameter)
 
     def run(self):
-        _LOGGER.info("Parameter Server Running!")
+        _LOGGER.info("DownpourSGD Running!")
         self.running = True
         while self.running:
             _LOGGER.info("Polling for data")
@@ -43,8 +40,8 @@ class ParameterServer():
             _LOGGER.info("Got message")
             self.receive(ACTION_CODES[self.m_parameter[0].item()], self.m_parameter[1:])
 
-
-
-
-
+def init_sgd():
+    model = Net()
+    server = DownpourSGD(learning_rate=DEFAULT_LEARNING_RATE, model=model)
+    server.run()
